backend/requirements.txt
Django==5.0.7
psycopg[binary]==3.2.10
djangorestframework==3.15.2
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
Pillow==11.0.0
qrcode==7.4.2
python-dotenv==1.0.1